# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/v1/user/password:
    put:
      tags:
      - user-rest-controller
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPasswordChangeRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/user/address:
    put:
      tags:
      - user-rest-controller
      operationId: saveAddress
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAddressUpdateRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - user-rest-controller
      operationId: deleteAddress
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/user/account:
    put:
      tags:
      - user-rest-controller
      operationId: updateAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAccountUpdateRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/auth/register:
    post:
      tags:
      - login-rest-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/auth/login:
    post:
      tags:
      - login-rest-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/v1/user:
    get:
      tags:
      - user-rest-controller
      operationId: getCurrentUser
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/home/products:
    get:
      tags:
      - home-rest-controller
      operationId: products
      parameters:
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HomeProductsResponseDTO"
  /api/v1/home/:
    get:
      tags:
      - home-rest-controller
      operationId: home
      parameters:
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: q
        in: query
        required: false
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HomeResponseDTO"
  /api/v1/home:
    get:
      tags:
      - home-rest-controller
      operationId: home_1
      parameters:
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: q
        in: query
        required: false
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HomeResponseDTO"
...
components:
  schemas:
    UserPasswordChangeRequestDTO:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
        confirmPassword:
          type: string
    UserAddressUpdateRequestDTO:
      type: object
      properties:
        street:
          type: string
        additional:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        country:
          type: string
        phone:
          type: string
    UserAccountUpdateRequestDTO:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        birthDate:
          type: string
...
    CategoryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        averageRating:
          type: number
          format: double
        category:
          $ref: "#/components/schemas/CategoryDTO"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ImageDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ProductReviewDTO"
    ProductReviewDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        rating:
          type: number
          format: double
        date:
          type: string
          format: date
        authorName:
          type: string
...
    RegisterRequestDTO:
      type: object
      properties:
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - FAILURE
        message:
          type: string
        error:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...
    HomeProductsResponseDTO:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        hasMore:
          type: boolean
    HomeResponseDTO:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryDTO"
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        selectedCategoryId:
          type: integer
          format: int64
        selectedCategoryName:
          type: string
        searchQuery:
          type: string
        hasMore:
          type: boolean
...